# extras/openapi.json
{
   "openapi":"3.1.0",
   "info":{
      "title":"Puddle Backend",
      "description":"\n    Backend API for Puddle, a dataset marketplace platform enabling\n    seamless interaction between dataset vendors and buyers, with AI agent\n    support for enhanced dataset processing and management. \n    Features include user authentication, dataset management, vector\n    similarity search, chat functionality, and comprehensive API\n    documentation.\n    ",
      "version":"1.0.0"
   },
   "paths":{
...
      "/api/v1/datasets/":{
         "post":{
            "tags":[
               "datasets"
            ],
            "summary":"Create Dataset",
            "operationId":"create_dataset_api_v1_datasets__post",
            "security":[
               {
                  "OAuth2PasswordBearer":[
                     
                  ]
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/DatasetCreate"
                     }
                  }
               }
            },
            "responses":{
               "201":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/DatasetRead"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "get":{
            "tags":[
               "datasets"
            ],
            "summary":"List Datasets",
            "operationId":"list_datasets_api_v1_datasets__get",
            "security":[
               {
                  "OAuth2PasswordBearer":[
                     
                  ]
               }
            ],
            "parameters":[
               {
                  "name":"search",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "anyOf":[
                        {
                           "type":"string"
                        },
                        {
                           "type":"null"
                        }
                     ],
                     "title":"Search"
                  }
               },
               {
                  "name":"limit",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"integer",
                     "maximum":500,
                     "minimum":1,
                     "default":50,
                     "title":"Limit"
                  }
               },
               {
                  "name":"offset",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"integer",
                     "minimum":0,
                     "default":0,
                     "title":"Offset"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/DatasetRead"
                           },
                           "title":"Response List Datasets Api V1 Datasets  Get"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/v1/datasets/{dataset_id}":{
         "get":{
            "tags":[
               "datasets"
            ],
            "summary":"Get Dataset",
            "operationId":"get_dataset_api_v1_datasets__dataset_id__get",
            "security":[
               {
                  "OAuth2PasswordBearer":[
                     
                  ]
               }
            ],
            "parameters":[
               {
                  "name":"dataset_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Dataset Id"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/DatasetRead"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "put":{
            "tags":[
               "datasets"
            ],
            "summary":"Update Dataset",
            "operationId":"update_dataset_api_v1_datasets__dataset_id__put",
            "security":[
               {
                  "OAuth2PasswordBearer":[
                     
                  ]
               }
            ],
            "parameters":[
               {
                  "name":"dataset_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Dataset Id"
                  }
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/DatasetUpdate"
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/DatasetRead"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "delete":{
            "tags":[
               "datasets"
            ],
            "summary":"Delete Dataset",
            "operationId":"delete_dataset_api_v1_datasets__dataset_id__delete",
            "security":[
               {
                  "OAuth2PasswordBearer":[
                     
                  ]
               }
            ],
            "parameters":[
               {
                  "name":"dataset_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Dataset Id"
                  }
               }
            ],
            "responses":{
               "204":{
                  "description":"Successful Response"
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/v1/datasets/search/embedding":{
         "post":{
            "tags":[
               "datasets"
            ],
            "summary":"Search By Embedding",
            "operationId":"search_by_embedding_api_v1_datasets_search_embedding_post",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/DatasetSearchQuery"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            },
            "security":[
               {
                  "OAuth2PasswordBearer":[
                     
                  ]
               }
            ]
         }
      },
...
   },
   "components":{
      "schemas":{
...
         "DatasetColumn":{
            "properties":{
               "name":{
                  "type":"string",
                  "title":"Name"
               },
               "description":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Description"
               },
               "data_type":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Data Type"
               },
               "sample_values":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Sample Values"
               }
            },
            "type":"object",
            "required":[
               "name"
            ],
            "title":"DatasetColumn"
         },
         "DatasetColumnRead":{
            "properties":{
               "name":{
                  "type":"string",
                  "title":"Name"
               },
               "description":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Description"
               },
               "data_type":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Data Type"
               },
               "sample_values":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Sample Values"
               },
               "id":{
                  "type":"integer",
                  "title":"Id"
               },
               "created_at":{
                  "type":"string",
                  "format":"date-time",
                  "title":"Created At"
               }
            },
            "type":"object",
            "required":[
               "name",
               "id",
               "created_at"
            ],
            "title":"DatasetColumnRead"
         },
         "DatasetCreate":{
            "properties":{
               "vendor_id":{
                  "type":"string",
                  "format":"uuid",
                  "title":"Vendor Id"
               },
               "title":{
                  "type":"string",
                  "title":"Title"
               },
               "description":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Description"
               },
               "domain":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Domain"
               },
               "dataset_type":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Dataset Type"
               },
               "granularity":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Granularity"
               },
               "pricing_model":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Pricing Model"
               },
               "license":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"License"
               },
               "topics":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Topics"
               },
               "entities":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Entities"
               },
               "temporal_coverage":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Temporal Coverage"
               },
               "geographic_coverage":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Geographic Coverage"
               },
               "visibility":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Visibility",
                  "default":"public"
               },
               "status":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Status",
                  "default":"active"
               },
               "columns":{
                  "anyOf":[
                     {
                        "items":{
                           "$ref":"#/components/schemas/DatasetColumn"
                        },
                        "type":"array"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Columns"
               },
               "embedding_input":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Embedding Input"
               },
               "embedding":{
                  "anyOf":[
                     {
                        "items":{
                           "type":"number"
                        },
                        "type":"array"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Embedding"
               }
            },
            "type":"object",
            "required":[
               "vendor_id",
               "title"
            ],
            "title":"DatasetCreate"
         },
         "DatasetRead":{
            "properties":{
               "vendor_id":{
                  "type":"string",
                  "format":"uuid",
                  "title":"Vendor Id"
               },
               "title":{
                  "type":"string",
                  "title":"Title"
               },
               "description":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Description"
               },
               "domain":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Domain"
               },
               "dataset_type":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Dataset Type"
               },
               "granularity":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Granularity"
               },
               "pricing_model":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Pricing Model"
               },
               "license":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"License"
               },
               "topics":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Topics"
               },
               "entities":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Entities"
               },
               "temporal_coverage":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Temporal Coverage"
               },
               "geographic_coverage":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Geographic Coverage"
               },
               "visibility":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Visibility",
                  "default":"public"
               },
               "status":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Status",
                  "default":"active"
               },
               "columns":{
                  "anyOf":[
                     {
                        "items":{
                           "$ref":"#/components/schemas/DatasetColumnRead"
                        },
                        "type":"array"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Columns"
               },
               "embedding_input":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Embedding Input"
               },
               "embedding":{
                  "anyOf":[
                     {
                        "items":{
                           "type":"number"
                        },
                        "type":"array"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Embedding"
               },
               "id":{
                  "type":"string",
                  "format":"uuid",
                  "title":"Id"
               },
               "created_at":{
                  "type":"string",
                  "format":"date-time",
                  "title":"Created At"
               },
               "updated_at":{
                  "type":"string",
                  "format":"date-time",
                  "title":"Updated At"
               }
            },
            "type":"object",
            "required":[
               "vendor_id",
               "title",
               "columns",
               "id",
               "created_at",
               "updated_at"
            ],
            "title":"DatasetRead"
         },
         "DatasetSearchQuery":{
            "properties":{
               "query":{
                  "type":"string",
                  "title":"Query"
               },
               "top_k":{
                  "type":"integer",
                  "title":"Top K",
                  "default":5
               }
            },
            "type":"object",
            "required":[
               "query"
            ],
            "title":"DatasetSearchQuery"
         },
         "DatasetUpdate":{
            "properties":{
               "title":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Title"
               },
               "description":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Description"
               },
               "domain":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Domain"
               },
               "dataset_type":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Dataset Type"
               },
               "granularity":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Granularity"
               },
               "pricing_model":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Pricing Model"
               },
               "license":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"License"
               },
               "topics":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Topics"
               },
               "entities":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Entities"
               },
               "temporal_coverage":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Temporal Coverage"
               },
               "geographic_coverage":{
                  "anyOf":[
                     {
                        
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Geographic Coverage"
               },
               "visibility":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Visibility"
               },
               "status":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Status"
               },
               "columns":{
                  "anyOf":[
                     {
                        "items":{
                           "$ref":"#/components/schemas/DatasetColumn"
                        },
                        "type":"array"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Columns"
               }
            },
            "type":"object",
            "title":"DatasetUpdate"
         },
         "HTTPValidationError":{
            "properties":{
               "detail":{
                  "items":{
                     "$ref":"#/components/schemas/ValidationError"
                  },
                  "type":"array",
                  "title":"Detail"
               }
            },
            "type":"object",
            "title":"HTTPValidationError"
         },
...
      },
      "securitySchemes":{
         "OAuth2PasswordBearer":{
            "type":"oauth2",
            "flows":{
               "password":{
                  "scopes":{
                     
                  },
                  "tokenUrl":"/api/v1/auth/login"
               }
            }
         }
      }
   },
...
}